create button follow child component to handle follow button on each individual card
